comment_service/requirements.txt
confluent-kafka==2.6.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.9.0
python-decouple==3.8
psycopg2==2.9.10
requests==2.32.3
